Consolidated all file helper functions into one file
