+++ src/events/ready/01registerCommands.js
@@ -1,7 +1,7 @@
 const { testServer } = require('../../../config.json');
-const getLocalCommands = require('../../utils/getLocalCommands');
-const getApplicationCommands = require('../../utils/getApplicationCommands');
-const areCommandsDifferent = require('../../utils/areCommandsDifferent');
+const GetLocalCommands = require("../utils/fileHelpers");
+const GetApplicationCommands = require("../utils/fileHelpers");
+const AreCommandsDifferent = require("../utils/fileHelpers");
 const { Client } = require('discord.js');
 
 /**
@@ -10,8 +10,8 @@ const { Client } = require('discord.js');
  */
 module.exports = async (client) => {
     try {
-        const localCommands = getLocalCommands();
-        const applicationCommands = await getApplicationCommands(client, testServer);
+        const localCommands = GetLocalCommands();
+        const applicationCommands = await GetApplicationCommands(client, testServer);
 
         // Delete a command by id
         // applicationCommands.delete('1094985345889738852');
@@ -35,7 +35,7 @@ module.exports = async (client) => {
                 }
 
                 // Check if local command is different
-                if (areCommandsDifferent(existingCommand, localCommand)) {
+                if (AreCommandsDifferent(existingCommand, localCommand)) {
                     await applicationCommands.edit(existingCommand.id, { description, options });
                     console.log(`Edited command "${name}".`);
                 }
